ai-service/requirements.txt
# AI Service - Final Clean Version
# Core AI/ML libraries
torch==2.1.1
torchvision==0.16.1
onnx==1.15.0
onnxruntime==1.16.3

# Computer Vision
opencv-python-headless==4.8.1.78
mediapipe==0.10.8
Pillow==10.1.0

# Data processing
numpy==1.24.4
pandas==2.1.4
scikit-learn==1.3.2

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# HTTP
httpx==0.25.2
requests==2.31.0

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0

# Monitoring
prometheus-client==0.19.0

# CORS
fastapi-cors==0.0.6

# Sentry SDK
sentry-sdk[fastapi]

# Gemini AI Integration
google-generativeai>=0.3.0
google-api-python-client>=2.100.0
